+++ event.go
@@ -26,8 +26,9 @@ const (
 )
 
 type Event struct {
-	ID      string         `json:"-"`
-	Seq     int            `json:"seq"`
-	Type    EventType      `json:"type"`
-	Payload map[string]any `json:"payload"`
+	ID        string         `json:"-"`
+	Seq       int            `json:"seq"`
+	Timestamp string         `json:"timestamp,omitempty"`
+	Type      EventType      `json:"type"`
+	Payload   map[string]any `json:"payload"`
 }